#87048 Add separate validator for PayloadTransform (#297)
