+++ src/CaptainHook.Application/Validators/Dtos/WebhooksValidatorDtoType.cs
@@ -1,0 +1,9 @@
+﻿namespace CaptainHook.Application.Validators.Dtos
+{
+    public enum WebhooksValidatorDtoType
+    {
+        Webhook = 0,
+        Callback,
+        DlqHook
+    }
+}
